# crackGreVocab/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Crack GRE Vocab API",
        "version": "0.1.0",
        "description": "Local API contract for the Milestone 1 cold rebuild."
    },
    "paths": {
...
        "/api/auth/account/": {
            "get": {
                "operationId": "auth_account_retrieve",
                "description": "Return the learner represented by the current server session.",
                "tags": [
                    "auth"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LearnerAccount"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/csrf/": {
            "get": {
                "operationId": "auth_csrf_retrieve",
                "description": "Issue the masked token required by session-backed mutations.",
                "tags": [
                    "auth"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CsrfToken"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/api/auth/password-reset/": {
            "post": {
                "operationId": "auth_password_reset_create",
                "description": "Accept a password recovery request.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-CSRFToken",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Masked token from GET /api/auth/csrf/.",
                        "required": true
                    }
                ],
                "tags": [
                    "auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetStartRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthValidationError"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/password-reset/confirm/": {
            "post": {
                "operationId": "auth_password_reset_confirm_create",
                "description": "Replace a password after validating one short-lived reset token.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-CSRFToken",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Masked token from GET /api/auth/csrf/.",
                        "required": true
                    }
                ],
                "tags": [
                    "auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetConfirmRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthValidationError"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/sign-in/": {
            "post": {
                "operationId": "auth_sign_in_create",
                "description": "Create a Django session from an email and password.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-CSRFToken",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Masked token from GET /api/auth/csrf/.",
                        "required": true
                    }
                ],
                "tags": [
                    "auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LearnerAccount"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthValidationError"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/sign-out/": {
            "post": {
                "operationId": "auth_sign_out_create",
                "description": "End the current browser session.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-CSRFToken",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Masked token from GET /api/auth/csrf/.",
                        "required": true
                    }
                ],
                "tags": [
                    "auth"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/auth/sign-up/": {
            "post": {
                "operationId": "auth_sign_up_create",
                "description": "Create and immediately sign in a new learner.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-CSRFToken",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Masked token from GET /api/auth/csrf/.",
                        "required": true
                    }
                ],
                "tags": [
                    "auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignUpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LearnerAccount"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthValidationError"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiMessage": {
                "type": "object",
                "description": "Describe an API response carrying one human-readable message.",
                "properties": {
                    "detail": {
                        "type": "string"
                    }
                },
                "required": [
                    "detail"
                ]
            },
            "AuthValidationError": {
                "type": "object",
                "description": "Describe malformed JSON or field-level account input errors.",
                "properties": {
                    "detail": {
                        "type": "string"
                    },
                    "email": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "display_name": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "password": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "token": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "uid": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "non_field_errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "CsrfToken": {
                "type": "object",
                "description": "Return a masked CSRF token for one unsafe request.",
                "properties": {
                    "csrf_token": {
                        "type": "string"
                    }
                },
                "required": [
                    "csrf_token"
                ]
            },
...
            "LearnerAccount": {
                "type": "object",
                "description": "Return the public portion of the current learner account.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "readOnly": true,
                        "title": "Email address"
                    },
                    "display_name": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "display_name",
                    "email",
                    "id"
                ]
            },
...
            "PasswordResetConfirmRequest": {
                "type": "object",
                "description": "Validate the opaque identity, token, and replacement password shape.",
                "properties": {
                    "uid": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 128
                    },
                    "token": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true,
                        "minLength": 1
                    }
                },
                "required": [
                    "password",
                    "token",
                    "uid"
                ]
            },
            "PasswordResetStartRequest": {
                "type": "object",
                "description": "Validate a non-enumerating password-recovery request.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1,
                        "maxLength": 254
                    }
                },
                "required": [
                    "email"
                ]
            },
...
            "SignInRequest": {
                "type": "object",
                "description": "Validate the credentials supplied for a new server session.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1,
                        "maxLength": 254
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true,
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "SignUpRequest": {
                "type": "object",
                "description": "Validate signup details and create a learner account.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1,
                        "maxLength": 254
                    },
                    "display_name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true,
                        "minLength": 1
                    }
                },
                "required": [
                    "display_name",
                    "email",
                    "password"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    }
}